refactor: hook으로 분리
