+++ frontend/src/pages/Admin/ModifyCouponPolicy/common/constant.ts
@@ -1,0 +1,6 @@
+export const MODIFY_STEP_NUMBER = {
+  createdType: 1,
+  maxStampCount: 2,
+  rewardName: 3,
+  expirePeriod: 4,
+};
