+++ frontend/src/pages/Admin/ModifyCouponPolicy/index.tsx
@@ -1,27 +1,18 @@
 import { useState } from 'react';
-import { useNavigate } from 'react-router-dom';
 import Button from '../../../components/Button';
 import { ButtonContainer, StepContainer, StepWrapper } from './style';
 import { Spacing } from '../../../style/layout/common';
 import Text from '../../../components/Text';
-import { EXPIRE_DATE_OPTIONS, ROUTER_PATH, STAMP_COUNT_OPTIONS } from '../../../constants';
+import { EXPIRE_DATE_OPTIONS, STAMP_COUNT_OPTIONS } from '../../../constants';
 import { CouponCreated, Option } from '../../../types';
 import CreatedType from './CreatedType';
 import MaxStampCount from './MaxStampCount';
 import ExpiredPeriod from './ExpirePeriod';
 import RewardName from './RewardName';
-import { selectRoutePathByCreatedType } from './logic';
-
-const MODIFY_STEP_NUMBER = {
-  createdType: 1,
-  maxStampCount: 2,
-  rewardName: 3,
-  expirePeriod: 4,
-};
+import { MODIFY_STEP_NUMBER } from './common/constant';
+import useStep from './hooks/useStep';
 
 const ModifyCouponPolicy = () => {
-  const navigate = useNavigate();
-  const [step, setStep] = useState(MODIFY_STEP_NUMBER.createdType);
   const [createdType, setCreatedType] = useState<CouponCreated>('template');
   const [rewardName, setRewardName] = useState('');
   const [expirePeriod, setExpirePeriod] = useState<Option>({
@@ -32,35 +23,12 @@ const ModifyCouponPolicy = () => {
     key: STAMP_COUNT_OPTIONS[0].key,
     value: STAMP_COUNT_OPTIONS[0].value,
   });
-
-  const navigateNextPage = () => {
-    navigate(`${ROUTER_PATH.modifyCouponPolicy}${selectRoutePathByCreatedType(createdType)}`, {
-      state: {
-        createdType,
-        reward: rewardName,
-        expirePeriod,
-        stampCount: stampCount.value,
-      },
-    });
-  };
-
-  const moveNextStep = () => {
-    if (step === MODIFY_STEP_NUMBER.rewardName && !rewardName) {
-      alert('리워드를 작성해주세요.');
-      return;
-    }
-
-    if (step === MODIFY_STEP_NUMBER.expirePeriod) {
-      navigateNextPage();
-      return;
-    }
-
-    setStep((prev) => (prev += 1));
-  };
-
-  const movePrevStep = () => {
-    setStep((prev) => (prev -= 1));
-  };
+  const { step, movePrevStep, moveNextStep } = useStep(
+    createdType,
+    rewardName,
+    expirePeriod,
+    stampCount.value,
+  );
 
   return (
     <>
